tests/wpt/web-platform-tests/cookies/prefix/__host.header.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/cookies/resources/cookie-helper.sub.js"></script>
<script>
  ["", "domain="+document.location.hostname, "MaxAge=10", "HttpOnly"].forEach(extraParams => {
    // Without 'secure'
    set_prefixed_cookie_via_http_test({
      prefix: "__Host-",
      params: "Path=/;" + extraParams,
      origin: self.origin,
      shouldExistInDOM: false,
      shouldExistViaHTTP: false,
      title: "__Host: Non-secure origin: Does not set 'Path=/;" + extraParams + "'"
    });

    // With 'secure'
    set_prefixed_cookie_via_http_test({
      prefix: "__Host-",
      params: "Secure; Path=/;" + extraParams,
      origin: self.origin,
      shouldExistInDOM: false,
      shouldExistViaHTTP: false,
      title: "__Host: Non-secure origin: Does not set 'Secure; Path=/;" + extraParams + "'"
    });

    // With 'domain'
    set_prefixed_cookie_via_http_test({
      prefix: "__Host-",
      params: "Secure; Path=/; Domain=" + document.location.hostname + "; " + extraParams,
      origin: self.origin,
      shouldExistInDOM: false,
      shouldExistViaHTTP: false,
      title: "__Host: Secure origin: Does not set 'Secure; Path=/; Domain=" + document.location.hostname + "; " + extraParams + "'"
    });
  });

  set_prefixed_cookie_via_http_test({
    prefix: "__Host-",
    params: "Secure; Path=/cookies/resources/list.py",
    origin: self.origin,
    shouldExistInDOM: false,
    shouldExistViaHTTP: false,
    title: "__Host: Non-secure origin: Does not set 'Secure; Path=/cookies/resources/list.py'"
  });
</script>
