LayoutTests/platform/glib/imported/w3c/web-platform-tests/cookies/prefix/__host.header-expected.txt


FAIL __Host: Non-secure origin: Does not set 'Path=/;' assert_equals: expected (undefined) undefined but got (string) "0.6673041091373431"
PASS __Host: Non-secure origin: Does not set 'Secure; Path=/;'
PASS __Host: Secure origin: Does not set 'Secure; Path=/; Domain=web-platform.test; '
FAIL __Host: Non-secure origin: Does not set 'Path=/;domain=web-platform.test' assert_equals: expected (undefined) undefined but got (string) "0.26675701551609077"
PASS __Host: Non-secure origin: Does not set 'Secure; Path=/;domain=web-platform.test'
PASS __Host: Secure origin: Does not set 'Secure; Path=/; Domain=web-platform.test; domain=web-platform.test'
FAIL __Host: Non-secure origin: Does not set 'Path=/;MaxAge=10' assert_equals: expected (undefined) undefined but got (string) "0.5687027309783798"
PASS __Host: Non-secure origin: Does not set 'Secure; Path=/;MaxAge=10'
PASS __Host: Secure origin: Does not set 'Secure; Path=/; Domain=web-platform.test; MaxAge=10'
FAIL __Host: Non-secure origin: Does not set 'Path=/;HttpOnly' assert_equals: expected (undefined) undefined but got (string) "0.148093504052612"
PASS __Host: Non-secure origin: Does not set 'Secure; Path=/;HttpOnly'
PASS __Host: Secure origin: Does not set 'Secure; Path=/; Domain=web-platform.test; HttpOnly'
PASS __Host: Non-secure origin: Does not set 'Secure; Path=/cookies/resources/list.py'
